fix: keep the agent-service recovery suite under the 1500-line guardrail
Commits ea893c9 and f99164 grew test_executor_agent_service_recovery.py to 1555
lines, tripping test_first_party_code_files_stay_under_line_limit alongside the
orchestrator-import failure. Move the four masked-timeout classification tests
those commits added into a sibling module that imports the shared fixtures,
following the pattern already used across tests/unit/control. No assertions
changed; the original file is back to 1451 lines.
